Splitting out defined classes
