+++ HydroModelBuilder/ModelInterface/flopyInterface/flopyInterface.py
@@ -1,6 +1,7 @@
 import datetime
 import os
 import warnings
+from types import MethodType
 
 import flopy
 import flopy.utils.binaryfile as bf
@@ -8,6 +9,8 @@ import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
 from flopy.utils.sfroutputfile import SfrFile
+
+from Radon_EC_simple import Radon_EC_simple
 
 
 class ModflowModel(object):
@@ -47,7 +50,7 @@ class ModflowModel(object):
         self.headtol = 1E-3  # 1E-6
         self.fluxtol = 1.0E4
 
-        if self.model_data.model_time.t['steady_state'] == True:
+        if self.model_data.model_time.t['steady_state']:
             self.nper = 1
             self.perlen = 1
             self.perlen = 40000 * 365
@@ -56,7 +59,7 @@ class ModflowModel(object):
         else:
             self.nper = self.model_data.model_time.t['steps']
             self.perlen = [
-                x.total_seconds() / 86400. for x in self.model_data.model_time.t['intervals']]
+                x.total_seconds() / 86400.0 for x in self.model_data.model_time.t['intervals']]
             self.nstp = 1
             self.steady = False
             self.start_datetime = self.model_data.model_time.t['dateindex'][0]
@@ -165,7 +168,7 @@ class ModflowModel(object):
     # End createSTRpackage()
 
     def createSFRpackage(self, reach_data, segment_data):
-        if self.model_data.model_time.t['steady_state'] == True:
+        if self.model_data.model_time.t['steady_state']:
             transroute = False
         else:
             transroute = True
@@ -337,8 +340,9 @@ class ModflowModel(object):
                 for key in bc_array.keys():
                     try:
                         river[key] += bc_array[key]
-                    except:
+                    except IndexError:
                         river[key] = bc_array[key]
+                    # End try
                 # End for
             # End if
 
@@ -1965,595 +1969,6 @@ class ModflowModel(object):
     # End viewGHB()
 
 # End ModflowModel()
-
-
-class MT3DModel(object):
-
-    def __init__(self, mf_model):
-        self.mf_model = mf_model
-
-    def createBTNpackage(self):
-        # Add the BTN package to the model
-        ibound = self.model_data.model_mesh3D[1]
-        ibound[ibound == -1] = 0
-        self.btn = flopy.mt3d.Mt3dBtn(self.mt, icbund=ibound, ncomp=1, mcomp=1,
-                                      cinact=-9.9E1, thkmin=-1.0E-6, ifmtcn=5,
-                                      ifmtnp=0, ifmtrf=0, ifmtdp=0, nprs=0,
-                                      timprs=None, savucn=1, nprobs=0,
-                                      chkmas=1, nprmas=1, dt0=10000.0, ttsmax=100000.0)
-
-    def createADVpackage(self):
-        # Add the ADV package to the model
-        self.adv = flopy.mt3d.Mt3dAdv(self.mt, mixelm=0, percel=1,
-                                      mxpart=250000, nadvfd=1, itrack=3,
-                                      wd=0.5, dceps=1.0E-4, nplane=0, npl=5,
-                                      nph=8, npmin=1, npmax=16, nlsink=0,
-                                      npsink=8, dchmoc=0.01)
-
-    def createDSPpackage(self):
-        # Add the DSP package to the model
-        self.dfp = flopy.mt3d.Mt3dDsp(self.mt, multiDiff=True, al=10.,
-                                      trpt=0.1, trpv=0.1, dmcoef=0.0)
-
-    def createRCTpackage(self):
-        # Add the RCT package to the model
-        self.rct = flopy.mt3d.Mt3dRct(self.mt, isothm=1, ireact=1,
-                                      igetsc=0, rc1=np.log(2) / (5730 * 365))
-
-    def createGCGpackage(self):
-        # Add the GCG package to the model
-        self.gcg = flopy.mt3d.Mt3dGcg(self.mt, mxiter=1000, iter1=100, isolve=1,
-                                      ncrs=0, accl=1, cclose=1.0E-4, iprgcg=0)
-
-    def finaliseMT3Dmodel(self):
-        warnings.warn("Deprecated. Use `finalize_MT3D_model()` instead.", DeprecationWarning)
-        self.finalize_MT3D_model()
-    # end finaliseMT3Dmodel
-
-    def finalize_MT3D_model(self):
-        self.mt.write_input()
-    # End finalize_MT3D_model()
-
-    def buildMT3D(self):
-
-        self.mt = flopy.mt3d.Mt3dms(modelname=self.name + '_transport', ftlfilename='mt3d_link.ftl',
-                                    modflowmodel=self.mf, model_ws=self.data_folder, exe_name='MT3D-USGS_64.exe')
-
-        self.createBTNpackage()
-        self.createADVpackage()
-        self.createDSPpackage()
-        self.createRCTpackage()
-        self.createGCGpackage()
-
-        ssm_data = {}
-        itype = flopy.mt3d.Mt3dSsm.itype_dict()
-        for per in range(self.nper):
-            ssm_data[per] = []
-        ibound = self.model_data.model_mesh3D[1]
-        ibound[ibound == -1] = 0
-
-        self.crch = {}
-        for per in range(self.nper):
-            self.crch[per] = []
-        # End for
-
-        river = {}
-        river[0] = []
-        bc = self.model_data.boundaries.bc
-        for boundary in bc:
-            bc_boundary = bc[boundary]
-            bc_type = bc_boundary['bc_type']
-            bc_array = bc_boundary['bc_array']
-
-            if (bc_type == 'river') or (bc_type == 'channel'):
-                time_key = bc_array.keys()[0]
-                river[0] += bc_array[time_key]
-            # End if
-
-            if bc_type == 'recharge':
-                for key in bc_array:
-                    self.crch[key] = np.ones_like(bc_array[key])
-                    self.crch[key] = self.crch[key] * 100.0
-                    self.crch[key][ibound[0] == 0] = 0.0
-                # End for
-            # End if
-
-            if bc_type == 'wells':
-                for key in bc_array:
-                    for well in bc_array[key]:
-                        ssm_data[key].append((well[0], well[1], well[2], 100.0, itype['WEL']))
-
-            if bc_type == 'general head':
-                for key in bc_array:
-                    for ghb in bc_array[key]:
-                        ssm_data[key].append((ghb[0], ghb[1], ghb[2], 0.0, itype['GHB']))
-                    # End for
-                # End for
-            # End if
-        # End for
-
-        if len(river) > 0:
-            for key in river:
-                for riv in river[key]:
-                    ssm_data[key].append((riv[0], riv[1], riv[2], 100.0, itype['RIV']))
-                # End for
-            # End for
-        # End if
-
-        self.ssm = flopy.mt3d.Mt3dSsm(self.mt, stress_period_data=ssm_data, crch=self.crch)
-        self.finaliseMT3Dmodel()
-    # End buildMT3D()
-
-    def runMT3D(self, silent=False):
-        success, buff = self.mt.run_model(silent=silent)
-        # if not success:
-        #    raise Exception('MT3D did not terminate normally.')
-        # End if
-    # End runMT3D()
-
-
-class Radon_EC_simple(object):
-
-    def __init__(self, df, Ini_cond, Rn_decay=0.181, units='m,d'):
-        self.Ini_Cond = Ini_cond  # 3 item list containing Initial flow, radon and ec concentrations
-        self.Rn_decay = Rn_decay  # constant for Radon decay
-        self.units = units  # units being used ... not really used yet!
-
-        # Dataframe variables
-        self.HZ_Poro = df['HZ_poro'].tolist()  # Hyporheic zone porosity
-        self.HZ_Prod_Rate = df['HZ_Prod_Rate'].tolist()  # Hyporheic zone radon production rate
-        self.HZ_RTime = df['HZ_RTime'].tolist()  # Hyporheic zone residence time
-        self.R_depth_HZ = df['R_depth_HZ'].tolist()
-        self.GTV = df['GTV'].tolist()  # Gas transfer velocity
-        self.GW_Rn_conc = df['GW_Rn_conc'].tolist()  # Groundwater radon concentration
-        self.GW_EC = df['GW_EC'].tolist()  # Groundwater EC
-        self.Trib_EC = df['Trib_EC'].tolist()
-        self.Trib_Rn = df['Trib_Rn'].tolist()
-        self.dx = df['dx'].tolist()
-        # Variables inherited from the output of SFR package
-        self.Evap_Rate = df['Qet'].tolist()  # Evaporation rate
-        self.Trib_inflow = df['Qovr'].tolist()
-
-        try:
-            self.R_Pump = df['R_pump'].tolist()
-        except:
-            self.R_Pump = [0.] * df.shape[0]
-        # end try
-        self.R_depth = df['depth'].tolist()
-        self.R_width = df['width'].tolist()
-        self.GW_inflow = (df['Qaquifer'] * -1.0).tolist()
-    # End __init__()
-
-    def Fl_Rn_EC_simul(self):
-
-        # Specify radon decay constant
-        Rn_decay = self.Rn_decay
-
-        # Initialise temp vars
-        Flow_temp = self.Ini_Cond[0]
-        Rn_temp = self.Ini_Cond[1]
-        EC_temp = self.Ini_Cond[2]
-        Fl_simul = [Flow_temp]
-        Rn_simul = [Rn_temp]
-        EC_simul = [EC_temp]
-
-        # Calculate simulated Radon and EC for the
-        for i in xrange(len(self.dx) - 1):
-            # Steady state stream flow calculation ... check against modelled
-            # as this won't consider rate of change in storage
-            Flow_temp += self.GW_inflow[i] - self.Evap_Rate[i] - \
-                self.R_Pump[i] + self.Trib_inflow[i]
-            # Radon concentration calculation based on steady state stream flow
-            if self.GW_inflow[i] < 0.0:
-                GW_Rn_conc_adj = Rn_temp
-            else:
-                GW_Rn_conc_adj = self.GW_Rn_conc[i]
-            # End if
-            Rn_temp += 1.0 / (Flow_temp) * (
-                self.GW_inflow[i] * (GW_Rn_conc_adj - Rn_temp) +
-                self.Trib_inflow[i] * (self.Trib_Rn[i] - Rn_temp) +
-                Rn_temp * (self.Evap_Rate[i] - self.dx[i] *
-                           self.R_width[i] * (self.GTV[i] + self.R_depth[i] * Rn_decay)) +
-                self.dx[i] * self.HZ_Poro[i] / (1 + Rn_decay * self.HZ_RTime[i]) *
-                self.R_width[i] * self.R_depth_HZ[i] * (self.HZ_Prod_Rate[i] - Rn_temp))
-            if Rn_temp < 0:
-                Rn_temp = 0.0
-            # EC calculation based on steady state stream flow
-            EC_temp += 1 / (Flow_temp) * (
-                self.GW_inflow[i] * (self.GW_EC[1] - EC_temp) +
-                self.Trib_inflow[i] * (self.Trib_EC[i] - EC_temp) +
-                EC_temp * self.Evap_Rate[i])
-
-            Fl_simul.append(Flow_temp)
-            Rn_simul.append(Rn_temp)
-            EC_simul.append(EC_temp)
-
-        return Fl_simul, Rn_simul, EC_simul
-    # End Fl_Rn_EC_simul()
-
-# End Radon_EC_simple()
-
-
-class MT3DPostProcess(object):
-
-    def __init__(self, mf_model, mt_name=None):
-        self.mf_model = mf_model
-        self.mt_name = mt_name
-
-    def importConcs(self):
-        self.concobj = bf.UcnFile(os.path.join(self.mf_model.data_folder, 'MT3D001.UCN'))
-        return self.concobj
-    # End importConcs()
-
-    def importSftConcs(self):
-        self.sft_conc = pd.read_csv(os.path.join(
-            self.mf_model.data_folder,
-            self.mt_name),
-            delim_whitespace=True,
-            skiprows=1)
-        return self.sft_conc
-    # End importSftConcs()
-
-    def ConcsByZone(self, concs):
-
-        self.mf_model.model_data.model_mesh3D[1]
-
-        concs_zoned = [np.full(self.mf_model.model_data.model_mesh3D[1].shape[
-                               1:3], np.nan)] * int(np.max(self.mf_model.model_data.model_mesh3D[1]))
-
-        for zone in range(int(np.max(self.mf_model.model_data.model_mesh3D[1]))):
-            temp_concs = np.array([np.full(self.mf_model.model_data.model_mesh3D[1].shape[
-                                  1:3], np.nan)] * self.mf_model.model_data.model_mesh3D[1].shape[0])
-            # for each layer in mesh get the heads from zone and average:
-            for layer in range(self.mf_model.model_data.model_mesh3D[1].shape[0]):
-                temp_concs[layer][self.mf_model.model_data.model_mesh3D[1][layer] == float(
-                    zone + 1)] = concs[layer][self.mf_model.model_data.model_mesh3D[1][layer] == float(zone + 1)]
-            masked_temp_concs = np.ma.masked_array(temp_concs, np.isnan(temp_concs))
-            concs_zoned[zone] = np.mean(masked_temp_concs, axis=0)
-        # End for
-        return concs_zoned
-    # End ConcsByZone()
-
-    def CompareObserved(self, obs_set, simulated, nper=0):
-
-        self.obs_sim_zone = []
-        obs_df = self.mf_model.model_data.observations.obs_group[obs_set]['time_series']
-        obs_df = obs_df[obs_df['active'] == True]
-        # obs_df = obs_df[obs_df['interval'] == nper]
-        sim_map_dict = self.mf_model.model_data.observations.obs_group[
-            obs_set]['mapped_observations']
-        # self.model_data.observations.obs_group[obs_set]['time_series']['name']:
-        for observation in obs_df['name']:
-            idx = obs_df[obs_df['name'] == observation].index.tolist()[0]
-            obs = obs_df.get_value(idx, 'value')
-            sim = simulated[sim_map_dict[observation][0]][
-                sim_map_dict[observation][1]][sim_map_dict[observation][2]]
-            zone = self.mf_model.model_data.model_mesh3D[1][sim_map_dict[observation][0]][
-                sim_map_dict[observation][1]][sim_map_dict[observation][2]]
-            x = self.mf_model.model_data.observations.obs_group[
-                obs_set]['locations']['Easting'].loc[observation]
-            y = self.mf_model.model_data.observations.obs_group[
-                obs_set]['locations']['Northing'].loc[observation]
-            if np.isnan(sim):
-                print sim, obs, zone
-                continue
-            self.obs_sim_zone += [[obs, sim, zone, x, y]]
-        # End for
-    # End CompareObserved()
-
-    def writeObservations(self, specimen):
-
-        # Set model output arrays to None to initialise
-        conc = None
-        sft_conc = None
-        data_folder = self.mf_model.data_folder
-        model_data = self.mf_model.model_data
-        obs_group = self.mf_model.model_data.observations.obs_group
-
-        # Write observation to file
-        for obs_set in model_data.observations.obs_group.keys():
-            obs_type = obs_group[obs_set]['obs_type']
-            # Import the required model outputs for processing
-            if obs_type not in ['concentration', 'EC', 'Radon']:
-                continue
-            else:
-                if (obs_type == 'concentration') & (specimen == 'C14'):
-                    # Check if model outputs have already been imported and if not import
-                    if not conc:
-                        concobj = self.importConcs()
-                        conc = concobj.get_alldata()  # (totim=times[0])
-                    # End if
-                elif (obs_type == 'EC') & (specimen == 'EC'):
-                    try:
-                        sft_conc['TIME']
-                    except:
-                        sft_conc = self.importSftConcs()
-                elif obs_type == 'Radon':
-                    continue
-                else:
-                    continue
-                # End if
-            # End if
-
-            obs_df = obs_group[obs_set]['time_series']
-            obs_df = obs_df[obs_df['active'] == True]
-            sim_map_dict = obs_group[obs_set]['mapped_observations']
-
-            with open(data_folder + os.path.sep + 'observations_' + obs_set + '.txt', 'w') as f:
-                if obs_group[obs_set]['domain'] == 'stream':
-                    sft_location = self.mf_model.model_data.observations.obs_group[obs_set]['locations']['seg_loc']
-
-                    for observation in obs_df.index:
-                        interval = int(obs_df['interval'].loc[observation])
-                        name = obs_df['name'].loc[observation]
-                        seg = sft_location.loc[name]
-                        sft = sft_conc
-                        times = sft['TIME'].unique()
-                        col_of_interest = obs_type
-                        if obs_type == 'EC':
-                            col_of_interest = 'SFR-CONCENTRATION'
-                        sim_conc = sft[(sft['SFR-NODE'] == seg) &
-                                       (sft['TIME'] == times[interval])][col_of_interest]
-                        f.write('%f\n' % sim_conc)
-
-                if obs_group[obs_set]['domain'] == 'porous':
-                    for observation in obs_df.index:
-                        interval = int(obs_df['interval'].loc[observation])
-                        name = obs_df['name'].loc[observation]
-
-                        (x_cell, y_cell) = model_data.mesh2centroid2Dindex[
-                            (sim_map_dict[name][1], sim_map_dict[name][2])]
-
-                        (lay, row, col) = [sim_map_dict[name][0],
-                                           sim_map_dict[name][1], sim_map_dict[name][2]]
-                        sim_conc = [conc[interval][lay][row][col]]
-                        sim_conc = np.mean(sim_conc)
-                        f.write('%f\n' % sim_conc)
-                    # End for
-                # End if
-            # End with
-        # End for
-    # End writeObservations()
-
-    def compareAllObs(self):
-
-        concobj = self.importConcs()
-        times = concobj.get_times()
-
-        scatterx = []
-        scattery = []
-        obs_sim_zone_all = []
-
-        # The definition of obs_sim_zone looks like:
-        for i in range(self.mf_model.model_data.model_time.t['steps']):
-            conc = concobj.get_data(totim=times[i])
-            self.CompareObserved('C14', conc, nper=i)
-            obs_sim_zone_all += self.obs_sim_zone
-
-        scatterx = np.array([h[0] for h in obs_sim_zone_all])
-        scattery = np.array([h[1] for h in obs_sim_zone_all])
-
-        # First step is to set up the plot
-        width = 20
-        height = 5
-        multiplier = 1.0
-        fig = plt.figure(figsize=(width * multiplier, height * multiplier))
-
-        ax = fig.add_subplot(1, 3, 1)
-        ax.set_title('Residuals')
-        ax.hist([loc[0] - loc[1] for loc in obs_sim_zone_all], bins=20, alpha=0.5)
-
-        ax = fig.add_subplot(1, 3, 2)
-        ax.set_title('Sim vs Obs (%d points)' % (len(scatterx)))
-
-        comp_zone_plots = {}
-        colours = ['r', 'orangered', 'y', 'green', 'teal', 'blue', 'fuchsia']
-        for i in xrange(1, 8):
-            scatterx2 = [loc[0] for loc in obs_sim_zone_all if loc[2] == float(i)]
-            scattery2 = [loc[1] for loc in obs_sim_zone_all if loc[2] == float(i)]
-            # print len(scatterx2), colours[i-1]
-            comp_zone_plots[i] = ax.scatter(scatterx2, scattery2, edgecolors=colours[
-                                            i - 1], facecolors='none', alpha=0.5)
-
-        plt.legend((comp_zone_plots[1], comp_zone_plots[2], comp_zone_plots[3],
-                    comp_zone_plots[4], comp_zone_plots[5], comp_zone_plots[6],
-                    comp_zone_plots[7]),
-                   ('qa', 'utb', 'utqa', 'utam', 'utaf', 'lta', 'bse'),
-                   scatterpoints=1,
-                   loc='upper left',
-                   ncol=4,
-                   fontsize=11)
-
-        plt.xlabel('Observed')
-        plt.ylabel('Simulated', labelpad=10)
-
-        sum1 = 0.0
-        sum2 = 0.0
-
-        if len(scatterx) != 0:
-
-            mean = np.mean(scatterx)
-            for i in range(len(scatterx)):
-                num1 = (scatterx[i] - scattery[i])
-                num2 = (scatterx[i] - mean)
-                sum1 += num1 ** np.float64(2.)
-                sum2 += num2 ** np.float64(2.)
-
-            ME = 1 - sum1 / sum2
-
-            ax.text(150, 75, 'Model Efficiency = %4.2f' % (ME))
-
-            # for PBIAS
-            def pbias(simulated, observed):
-                return np.sum(simulated - observed) * 100 / np.sum(observed)
-
-            ax.text(150, 40, 'PBIAS = %4.2f%%' % (pbias(scattery, scatterx)))
-
-            # For rmse
-            def rmse(simulated, observed):
-                return np.sqrt(((simulated - observed) ** 2).mean())
-
-            ax.text(150, 20, 'RMSE = %4.2f' % (rmse(scattery, scatterx)))
-
-        ax.plot(ax.get_ylim(), ax.get_ylim())
-
-        ax = fig.add_subplot(1, 3, 3)
-        ax.set_title('Residuals in space')
-
-        modelmap = flopy.plot.ModelMap(model=self.mf_model.mf)
-        modelmap.plot_ibound()
-
-        x = np.array([h[3] for h in obs_sim_zone_all])
-        y = np.array([h[4] for h in obs_sim_zone_all])
-        zone = [h[2] for h in obs_sim_zone_all]
-        residuals = [h[0] - h[1] for h in obs_sim_zone_all]
-        residuals = np.absolute(residuals)
-
-        from matplotlib import colors
-        import six
-        colors_ = list(six.iteritems(colors.cnames))
-        # Add the single letter colors.
-        for name, rgb in six.iteritems(colors.ColorConverter.colors):
-            hex_ = colors.rgb2hex(rgb)
-            colors_.append((name, hex_))
-
-        hex_ = [color[1] for color in colors_]
-        nams = [color[0] for color in colors_]
-        # Get the rgb equivalent.
-        rgb_all = [colors.hex2color(color) for color in hex_]
-
-        rgb_ref = []
-        for col in colours:
-            for index, nam in enumerate(nams):
-                if col == nam:
-                    rgb_ref += [rgb_all[index]]
-                # End if
-            # End for
-        # End for
-
-        zone = np.array(zone)
-        rgba_colors = np.zeros((len(x), 4))
-        # for red the first column needs to be one
-        for i in range(1, 8):
-            rgba_colors[:, 0][zone == i] = rgb_ref[i - 1][0]
-            rgba_colors[:, 1][zone == i] = rgb_ref[i - 1][1]
-            rgba_colors[:, 2][zone == i] = rgb_ref[i - 1][2]
-        # the fourth column needs to be your alphas
-        rgba_colors[:, 3] = residuals / np.max(residuals)  # alphas
-
-        plt.scatter(x, y, color=rgba_colors)
-        plt.show()
-    # End compareAllObs()
-
-    def viewConcsByZone(self, nper='all', specimen=None):
-
-        # Create the headfile object
-        concobj = self.importConcs()
-        times = concobj.get_times()
-        if nper == 'all':
-            conc = concobj.get_alldata()
-            conc = np.mean(conc, axis=0)
-            zoned = self.ConcsByZone(conc)
-            conc = zoned
-        elif nper == 'final':
-            conc = concobj.get_data(totim=times[-1])
-            zoned = self.ConcsByZone(conc)
-            conc = zoned
-        else:
-            conc = concobj.get_data(totim=times[nper])
-            zoned = self.ConcsByZone(conc)
-            conc = zoned
-        # End if
-
-        # First step is to set up the plot
-        width = 20
-        height = 10
-        multiplier = 1.
-        fig = plt.figure(figsize=(width * multiplier, height * multiplier))
-
-        vmin = np.amin(conc[conc > 0.])  # 0.0
-        vmax = np.amax(conc)  # 100.0
-
-        ax = fig.add_subplot(2, 4, 1, aspect='equal')
-
-        ax.set_title('ibound and bc')
-        # Next we create an instance of the ModelMap class
-        modelmap = flopy.plot.ModelMap(model=self.mf_model.mf)
-        modelmap.plot_ibound()
-
-        modelmap.plot_bc('RIV', plotAll=True)
-        modelmap.plot_bc('WEL', plotAll=True)
-        modelmap.plot_bc('GHB', plotAll=True)
-        modelmap.plot_bc('SFR', plotAll=True)
-        modelmap.plot_bc('DRN', plotAll=True)
-        ax.axes.xaxis.set_ticklabels([])
-
-        ax = fig.add_subplot(2, 4, 2, aspect='equal')
-        ax.set_title('Coonambidgal')
-        modelmap = flopy.plot.ModelMap(model=self.mf_model.mf)
-        max_conc = -100.0
-        min_conc = 1E6
-
-        array = modelmap.plot_array(
-            conc[0], masked_values=[-999.98999023, max_conc, min_conc], alpha=0.5, vmin=vmin, vmax=vmax)
-
-        ax.xaxis.set_ticklabels([])
-        ax.yaxis.set_ticklabels([])
-        cbar_ax2 = fig.add_axes([0.43, 0.525, 0.01, 0.42])
-        fig.colorbar(array, cax=cbar_ax2)
-
-        ax = fig.add_subplot(2, 4, 3, aspect='equal')
-        ax.set_title('Shepparton')
-        modelmap = flopy.plot.ModelMap(model=self.mf_model.mf)
-        array = modelmap.plot_array(
-            conc[2], masked_values=[-999.98999023, max_conc, min_conc, np.nan], alpha=0.5, vmin=vmin, vmax=vmax)
-        ax.xaxis.set_ticklabels([])
-        ax.yaxis.set_ticklabels([])
-        cbar_ax1 = fig.add_axes([0.67, 0.525, 0.01, 0.42])
-        fig.colorbar(array, cax=cbar_ax1)
-
-        ax = fig.add_subplot(2, 4, 5, aspect='equal')
-        ax.set_title('Calivil')
-        # , sr=self.mf.dis.sr, dis=self.mf.dis)
-        modelmap = flopy.plot.ModelMap(model=self.mf_model.mf)
-        array = modelmap.plot_array(
-            conc[4], masked_values=[-999.98999023, max_conc, min_conc, np.nan], alpha=0.5, vmin=vmin, vmax=vmax)
-        start, end = ax.get_xlim()
-        start = start // 1000 * 1000 + 1000
-        end = end // 1000 * 1000 - 1000
-        ax.xaxis.set_ticks(np.arange(start, end, 20000.))
-        ax.xaxis.get_major_formatter().set_powerlimits((0, 1))
-        cbar_ax3 = fig.add_axes([0.19, 0.055, 0.01, 0.42])
-        fig.colorbar(array, cax=cbar_ax3)
-
-        ax = fig.add_subplot(2, 4, 6, aspect='equal')
-        ax.set_title('Renmark')
-        # , sr=self.mf.dis.sr, dis=self.mf.dis)
-        modelmap = flopy.plot.ModelMap(model=self.mf_model.mf)
-        array = modelmap.plot_array(
-            conc[5], masked_values=[-999.98999023, max_conc, min_conc, np.nan], alpha=0.5, vmin=vmin, vmax=vmax)
-        ax.yaxis.set_ticklabels([])
-        ax.xaxis.set_ticks(np.arange(start, end, 20000.))
-        ax.xaxis.get_major_formatter().set_powerlimits((0, 1))
-        cbar_ax4 = fig.add_axes([0.43, 0.055, 0.01, 0.42])
-        fig.colorbar(array, cax=cbar_ax4)
-
-        ax = fig.add_subplot(2, 4, 7, aspect='equal')
-        ax.set_title('Basement')
-        modelmap = flopy.plot.ModelMap(model=self.mf_model.mf)
-        array = modelmap.plot_array(
-            conc[6], masked_values=[-999.98999023, max_conc, min_conc, np.nan], alpha=0.5, vmin=vmin, vmax=vmax)
-        ax.yaxis.set_ticklabels([])
-        ax.xaxis.set_ticks(np.arange(start, end, 20000.0))
-        ax.xaxis.get_major_formatter().set_powerlimits((0, 1))
-        cbar_ax5 = fig.add_axes([0.67, 0.055, 0.01, 0.42])
-        fig.colorbar(array, cax=cbar_ax5)
-        fig.subplots_adjust(left=0.01, right=0.95, bottom=0.05, top=0.95, wspace=0.1, hspace=0.12)
-
-        plt.show()
-    # End viewConcsByZone()
-
-# End MT3DPostProcess()
 
 
 if __name__ == '__main__':
